+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 
 ## Unreleased
 
+* Added macro feature to Toolproof
+
 ## v0.5.0 (November 28, 2024)
 
 * Add `before_all` commands to the Toolproof config